kconfig: enable moving files cross-device links
Issue: It is not possible to build inside a docker container with a mounted
folder for the output. Due to the usage of 'rename()' in kconfig the moving of
initial configs will fail early with errno: cross-device link.

Fix: Proposed is simple a move implementation. The approach shall keep the same
function signature and same behavior, i.e. existing target files will be
overwritten; fail at missing sources. Minimal for readability. No fancy
implementation.